+++ nl.vu_compmedchem.klifs/src/nl/vu_compmedchem/klifs/structures/StructuresGetMOL2NodeModel.java
@@ -21,8 +21,8 @@ import org.knime.core.node.BufferedDataTable;
 import org.knime.core.node.CanceledExecutionException;
 import org.knime.core.node.defaultnodesettings.SettingsModelString;
 
-import io.swagger.client.ApiException;
-import io.swagger.client.api.StructuresApi;
+import org.openapitools.client.ApiException;
+import org.openapitools.client.api.StructuresApi;
 import nl.vu_compmedchem.klifs.KlifsNodeModel;
 
 import org.knime.core.node.ExecutionContext;